MultiModal/requirements.txt
flask>=3.0.3
python-dotenv>=1.0.0
pillow>=10.3.0

unstructured[all-docs]==0.17.0  # Pinned version that supports these parameters
pypdf>=4.2.0

google-generativeai>=0.5.0
langchain-google-genai>=0.1.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.4.22

sentence-transformers>=3.0.0
huggingface-hub>=0.23.0

tiktoken>=0.7.0

pypdf==4.2.0
python-magic==0.4.27 